Add additional API routes for organizations CRUD actions
